+++ src/helpers/sortBuilder.tsx
@@ -5,7 +5,6 @@ interface SortBy<T> {
     descending: SortFn<T>
 }
 
-const isKey = <T, K extends keyof T, U>(f: K | U): f is K => typeof f === 'string'
 export const contramap = <T, K>(f: (t: T) => K, compareFn: SortFn<K> = (a: any, b: any) => a - b): SortFn<T> => (a2, b2) => compareFn(f(a2), f(b2))
 
 export const sortBy = <T, >() => <K extends keyof T>(field: K, compareFn: SortFn<T[K]> = (a: any, b: any) => a - b): SortBy<T> => ({
